# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API",
        "contact": {
            "name": "fersaavedra85@hotmail.com",
            "url": "https://abostudio.mx"
        },
        "version": "1.0"
    },
    "host": "develop.com",
    "basePath": "/ruffless/public/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "api/dashboard/group": {
            "get": {
                "tags": [
                    "grupos"
                ],
                "summary": "getGroup",
                "description": "Regresa los grupos a los que le usuario pertenece , asi como los amigos confirmados .",
                "operationId": "api.dashboard.getGroup",
                "consumes": [
                    "application/xml",
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID del usuario",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login success."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "post": {
                "tags": [
                    "grupos"
                ],
                "summary": "postGroup",
                "description": "Regresa el ranquin, general individual y por quincena.",
                "operationId": "api.dashboard.postGroup",
                "consumes": [
                    "application/xml",
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID del usuario",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "ID del usuario",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login success."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "delete": {
                "tags": [
                    "grupos"
                ],
                "summary": "deleteGroup",
                "description": "Regresa el ranquin, general individual y por quincena.",
                "operationId": "api.dashboard.deleteGroup",
                "consumes": [
                    "application/xml",
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID del usuario",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "group_id del grupo que se desea eliminar",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login success."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
...
    },
...
}